+++ packages/jason-api/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add `queryCache` option to control when `useRequest()`
+  and `useAutoRequest()` query the cache.
+
+## [1.1.1]
 ### Changed
 - Dependency list is no longer need for hooks, and `propsToWatch`
   is no longer necessary for High Order Components.